+++ src/utils/copyExample.ts
