+++ src/utils/promptForExample.ts
